requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
# awsebcli==3.20.3
backports.zoneinfo; python_version<"3.9"
bcrypt==4.1.2
bleach==6.0.0
blessed==1.20.0
botocore==1.23.54
cached-property==1.5.2
cachetools==5.3.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==41.0.3
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
Django==4.1.5
django-category==2.0.1
django-crispy-forms==2.0
django-email-verification==0.3.2
django-ipware==5.0.0
django-object-tools==2.0.0
django-recaptcha==3.0.0
django-summernote==0.8.20.0
django-tinymce==3.6.1
django-user-agents==0.4.0
djangorestframework==3.14.0
# docker==4.4.4
# docker-compose==1.25.5
# dockerpty==0.4.1
docopt==0.6.2
filelock==3.9.0
frozenlist==1.4.1
future==0.16.0
idna==3.4
ipinfo==4.2.1
ipinfo-django==2.0.0
jmespath==0.10.0
jsonschema==3.2.0
multidict==6.0.4
# mysqlclient==2.0.3
npm==0.1.1
oauthlib==3.2.2
optional-django==0.1.0
packaging==23.1
paramiko==3.4.0
pathspec==0.9.0
Pillow==10.0.0
platformdirs==2.6.2
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyrsistent==0.20.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
six==1.14.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.3
termcolor==1.1.0
texttable==1.7.0
typing-extensions==4.9.0
tzdata==2022.7
ua-parser==0.18.0
urllib3==1.26.14
user-agents==2.2.0
vercel==0.2.1
virtualenv==20.17.1
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
yarl==1.9.4
